Moved `AuthProcessable` and `AuthDelegate` to their own file
