+++ CareForMe/Features/Auth/View/AuthView.swift
@@ -7,32 +7,6 @@
 
 import UIKit
 
-enum AuthType {
-    case registration
-    case login
-}
-
-protocol AuthProcessable: AnyObject {
-    var loginDelegate: LoginProcessable? { get }
-    var registrationDelegate: RegistrationProcessable? { get }
-    init(_ delegate: LoginProcessable)
-    init(_ delegate: RegistrationProcessable)
-}
-
-class AuthDelegate: AuthProcessable {
-    
-    weak var loginDelegate: LoginProcessable?
-    weak var registrationDelegate: RegistrationProcessable?
-    
-    required init(_ delegate: RegistrationProcessable) {
-        self.registrationDelegate = delegate
-    }
-    
-    required init(_ delegate: LoginProcessable) {
-        self.loginDelegate = delegate
-    }
-}
-
 protocol RegistrationProcessable: AnyObject {
     func processRegistration(email: String, password: String)
 }
